requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.1.13
django-environ==0.10.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
drf-spectacular==0.26.5
environ==1.0
flake8==6.1.0
inflection==0.5.1
isort==5.12.0
jsonfield==3.1.0
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
pymongo==3.12.3
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.12.0
sqlparse==0.2.4
uritemplate==4.1.1
